[Feature] Renaming 교과목조회
