+++ src/components/views/ClassSearchTable.js
